chore: rename TouristDashboard.vue to TouristDashboardView.vue
