+++ src/router/index.ts
@@ -7,7 +7,7 @@ import SignUpView from '@/views/SignUp/SignUpView.vue';
 import HomeView from '@/views/Home/HomeView.vue';
 import ExperiencesView from '@/views/Experiences/ExperiencesView.vue';
 import TestingComponents from '@/views/testing-components/TestingComponents.vue';
-import TouristDashboard from '@/views/TouristDashboard/TouristDashboard.vue';
+import TouristDashboard from '@/views/TouristDashboard/TouristDashboardView.vue';
 
 const auth = getAuth();
