+++ src/views/TouristDashboard/TouristDashboardView.vue
